Фикс первой ошибки при входе

+++ server/include/NetworkModule.h
@@ -32,7 +32,7 @@ public:
 	uint8_t appendClient();												// добавляем клиента в отслеживаемый массив !узнать тип event
 	uint8_t removeClient(int32_t client_socket);						// убрать клиента из списка отслеживаемых
 	uint8_t getMessage(int32_t client_socket, char *buffer);			// получить сообщение
-	uint8_t sendMessage(int32_t client_socket, const char *buffer);		// отправить сообщение
+	int sendMessage(int32_t client_socket, const char *buffer);			// отправить сообщение
 	pollfd *readyFd();													// возвращает указатель на готовый к обработке файловый дескриптор
 	uint8_t getClientSocket(uint32_t client_uid, int32_t *dest_socket); // возвращает сокет клиента или ошибку
 	pollfd *getFd(unsigned int index);									// возвращает указатель на элемент fds[index]
